+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unreleased
 * BraintreeThreeDSecure
-  * Add `cardAddChallengeRequested` to `BTThreeDSecureRequest`
+  * Add `cardAddChallengeRequested`, `uiType`, and `renderTypes` to `BTThreeDSecureRequest`
   * Deprecate `BTThreeDSecureRequest.cardAddChallenge`
   * Fix bug where defaults for `BTThreeDSecureRequest.accountType`, `BTThreeDSecureRequest.requestedExemptionType`, and `BTThreeDSecureRequest.dfReferenceID` were improperly returning an error if not passed into the request
 * BraintreeCard